requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.5
pytest-django==4.10.0
requests==2.32.3
requests-mock==1.12.1
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
